requirements.txt
# LangChain Core Dependencies
langchain==0.3.7
langchain-core==0.3.15
langchain-google-genai==2.0.5

# Environment Variables
python-dotenv==1.0.0

# Numerical Computing (for embeddings examples)
numpy==1.26.4

# Optional: For better output formatting
colorama==0.4.6
